Beta

+++ py/ztools/mtp/mtpxci.py
@@ -4,11 +4,11 @@ import os
 import shutil
 import json
 import listmanager
-# from tqdm import tqdm
 from Fs import Nsp as squirrelNSP
 from Fs import Xci as squirrelXCI
 from Fs import factory
 from Fs.Nca import NcaHeader
+from Fs import Nca
 from Fs.File import MemoryFile
 from Fs import Ticket
 import sq_tools
@@ -24,6 +24,7 @@ from mtp.wpd import is_switch_connected
 from python_pick import pick
 from python_pick import Picker
 import csv
+from tqdm import tqdm
 
 if not is_switch_connected():
 	sys.exit("Switch device isn't connected.\nCheck if mtp responder is running!!!")	
@@ -247,53 +248,41 @@ def generate_xci_and_transfer(filepath=None,outfolder=None,destiny="SD",kgpatch=
 	keypatch=int(tgkg)	
 	tname=str(os.path.basename(filepath))[:-3]+'xci'
 	tmpfile=os.path.join(outfolder,tname)	
-	if isExe==False:
-		process0=subprocess.Popen([sys.executable,squirrel,"-lib_call","mtp.mtpxci","generate_and_transfer_st1","-xarg",filepath,outfolder,str(keypatch)])	
-	else:
-		process0=subprocess.Popen([squirrel,"-lib_call","mtp.mtpxci","generate_and_transfer_st1","-xarg",filepath,outfolder,str(keypatch)])		
-	while process0.poll()==None:
-		if process0.poll()!=None:
-			process0.terminate();
-	if isExe==False:
-		process1=subprocess.Popen([sys.executable,squirrel,"-renf",tmpfile,"-t","xci","-renm","force","-nover","xci_no_v0","-addl","false","-roma","TRUE"])
-	else:	
-		process1=subprocess.Popen([sys.executable,squirrel,"-renf",tmpfile,"-t","xci","-renm","force","-nover","xci_no_v0","-addl","false","-roma","TRUE"])		
-	while process1.poll()==None:
-		if process1.poll()!=None:
-			process1.terminate();	
-	files2transfer=listmanager.folder_to_list(outfolder,['xci'])
-	for f in files2transfer:
-		bname=str(os.path.basename(f))
-		destinypath=os.path.join(destiny,bname)
-		process=subprocess.Popen([nscb_mtp,"Transfer","-ori",f,"-dst",destinypath])		
-		while process.poll()==None:
-			if process.poll()!=None:
-				process.terminate();	
-	try:			
-		for f in os.listdir(outfolder):
-			fp = os.path.join(outfolder, f)
-			try:
-				shutil.rmtree(fp)
-			except OSError:
-				os.remove(fp)	
-	except:pass			
+	if filepath.endswith('xcz') or filepath.endswith('nsz'):
+		if isExe==False:
+			process0=subprocess.Popen([sys.executable,squirrel,"-lib_call","mtp.mtpxci","generate_and_transfer_st1","-xarg",filepath,outfolder,str(keypatch)])	
+		else:
+			process0=subprocess.Popen([squirrel,"-lib_call","mtp.mtpxci","generate_and_transfer_st1","-xarg",filepath,outfolder,str(keypatch)])		
+		while process0.poll()==None:
+			if process0.poll()!=None:
+				process0.terminate();
+		if isExe==False:
+			process1=subprocess.Popen([sys.executable,squirrel,"-renf",tmpfile,"-t","xci","-renm","force","-nover","xci_no_v0","-addl","false","-roma","TRUE"])
+		else:	
+			process1=subprocess.Popen([sys.executable,squirrel,"-renf",tmpfile,"-t","xci","-renm","force","-nover","xci_no_v0","-addl","false","-roma","TRUE"])		
+		while process1.poll()==None:
+			if process1.poll()!=None:
+				process1.terminate();	
+		files2transfer=listmanager.folder_to_list(outfolder,['xci'])
+		for f in files2transfer:
+			bname=str(os.path.basename(f))
+			destinypath=os.path.join(destiny,bname)
+			process=subprocess.Popen([nscb_mtp,"Transfer","-ori",f,"-dst",destinypath])		
+			while process.poll()==None:
+				if process.poll()!=None:
+					process.terminate();	
+		try:			
+			for f in os.listdir(outfolder):
+				fp = os.path.join(outfolder, f)
+				try:
+					shutil.rmtree(fp)
+				except OSError:
+					os.remove(fp)	
+		except:pass
+	elif filepath.endswith('xci') or filepath.endswith('nsp'):		
+		from mtpxci_gen import transfer_xci_csv
+		transfer_xci_csv(filepath,destiny,cachefolder=outfolder,keypatch=keypatch)	
 	
-	
-# def generate_multixci_and_transfer_st1(filepath,outfolder,keypatch='false'):
-	# tname=str(os.path.basename(filepath))[:-3]+'xci'
-	# tmpfile=os.path.join(outfolder,tname)
-	# if filepath.endswith('xci'):
-		# f = factory(filepath)
-		# f.open(filepath, 'rb')
-	# elif filepath.endswith('nsp'):	
-		# f = squirrelNSP(filepath, 'rb')	
-	# f.c_xci_direct(65536,tmpfile,outfolder,metapatch='true',keypatch=keypatch)
-	# f.flush()
-	# f.close()		
-
-# def generate_xci_and_transfer(filepath=None,outfolder=None,destiny="SD",kgpatch=False,verification=False):
-
-
 def generate_multixci_and_transfer(tfile=None,outfolder=None,destiny="SD",kgpatch=False,verification=False):
 	if destiny==False or destiny=="pick" or destiny=="":
 		destiny=pick_transfer_folder()	
@@ -500,7 +489,6 @@ def get_header_size(filepath):
 	return properheadsize,keygeneration,sz
 	
 def install_xci_csv(filepath,destiny="SD",cachefolder=None,override=False,keypatch=False):
-	import time
 	if cachefolder==None:
 		cachefolder=os.path.join(ztools_dir, '_mtp_cache_')	
 	files_list=sq_tools.ret_xci_offsets(filepath)
@@ -517,18 +505,27 @@ def install_xci_csv(filepath,destiny="SD",cachefolder=None,override=False,keypat
 		cnmtfile=entry[0]	
 		if cnmtfile.endswith('.cnmt.nca'):
 			target_cnmt=cnmtfile
-			nspname=gen_xci_parts(filepath,target_cnmt=target_cnmt,cachefolder=cachefolder,keypatch=keypatch,export_type='csv')
+			nspname=gen_xci_parts_spec1(filepath,target_cnmt=target_cnmt,cachefolder=cachefolder,keypatch=keypatch)
+			if filepath.endswith('xcz'):
+				nspname=nspname[:-1]+'z'
 			files_csv=os.path.join(cachefolder, 'files.csv')	
 			process=subprocess.Popen([nscb_mtp,"InstallfromCSV","-cs",files_csv,"-nm",nspname,"-dst",destiny])		
 			while process.poll()==None:
 				if process.poll()!=None:
 					process.terminate();	
 			counter-=1		
-			print('\n...................................................')
-			print('STILL '+str(counter)+' FILES TO PROCESS')
-			print('...................................................\n')		
+			print('\n- Still '+str(counter)+' subitems to process')	
+			if counter>0:
+				print("")			
+	if os.path.exists(cachefolder):			
+		for f in os.listdir(cachefolder):
+			fp = os.path.join(cachefolder, f)
+			try:
+				shutil.rmtree(fp)
+			except OSError:
+				os.remove(fp)			
 
-def gen_xci_parts(filepath,target_cnmt=None,cachefolder=None,keypatch=False,export_type='csv'):
+def gen_xci_parts_spec0(filepath,target_cnmt=None,cachefolder=None,keypatch=False,export_type='csv'):
 	if cachefolder==None:
 		cachefolder=os.path.join(ztools_dir, '_mtp_cache_')	
 	if not os.path.exists(cachefolder):
@@ -569,8 +566,16 @@ def gen_xci_parts(filepath,target_cnmt=None,cachefolder=None,keypatch=False,expo
 					test1=str(row['NcaId'])+'.nca';test2=str(row['NcaId'])+'.ncz'
 					if test1 in fplist or test2 in fplist:
 						# print(str(row['NcaId'])+'.nca')
-						files.append(str(row['NcaId'])+'.nca')
-						filesizes.append(int(row['Size']))	
+						if test1 in fplist:
+							files.append(str(row['NcaId'])+'.nca')
+							filesizes.append(int(row['Size']))
+						elif test2 in fplist:	
+							files.append(str(row['NcaId'])+'.ncz')
+							for k in range(len(files_list)):
+								entry=files_list[k]
+								if entry[0]==test2:				
+									filesizes.append(int(entry[3]))	
+									break
 			for j in range(len(ncadata)):
 				row=ncadata[j]						
 				if row['NCAtype']=='Meta':
@@ -584,12 +589,16 @@ def gen_xci_parts(filepath,target_cnmt=None,cachefolder=None,keypatch=False,expo
 					test1=str(row['NcaId'])+'.nca';test2=str(row['NcaId'])+'.ncz'
 					if test1 in fplist or test2 in fplist:
 						# print(str(row['NcaId'])+'.nca')
-						files.append(str(row['NcaId'])+'.nca')
-						filesizes.append(int(row['Size']))
 						if test1 in fplist:
-							program_name=test1
+							files.append(str(row['NcaId'])+'.nca')
+							filesizes.append(int(row['Size']))
 						elif test2 in fplist:	
-							program_name=test2
+							files.append(str(row['NcaId'])+'.ncz')
+							for k in range(len(files_list)):
+								entry=files_list[k]
+								if entry[0]==test2:				
+									filesizes.append(int(entry[3]))	
+									break
 			break			
 	f.flush()
 	f.close()						
@@ -633,6 +642,7 @@ def gen_xci_parts(filepath,target_cnmt=None,cachefolder=None,keypatch=False,expo
 			if str(nspF._path)=="secure":
 				for nca in nspF:	
 					if nca._path==fi:
+						nca=Nca(nca)
 						crypto1=nca.header.getCryptoType()
 						crypto2=nca.header.getCryptoType2()	
 						if crypto2>crypto1:
@@ -653,14 +663,18 @@ def gen_xci_parts(filepath,target_cnmt=None,cachefolder=None,keypatch=False,expo
 							titleKeyDec = Keys.decryptTitleKey(titleKey, Keys.getMasterKeyIndex(int(masterKeyRev)))							
 							encKeyBlock = crypto.encrypt(titleKeyDec * 4)
 							if str(keypatch) != "False":
+								t = tqdm(total=False, unit='B', unit_scale=False, leave=False)								
 								if keypatch < nca.header.getCryptoType2():
-									encKeyBlock,crypto1,crypto2=squirrelXCI.get_new_cryptoblock(nca, keypatch,encKeyBlock,t)	
+									encKeyBlock,crypto1,crypto2=squirrelXCI.get_new_cryptoblock(squirrelXCI,nca, keypatch,encKeyBlock,t)	
+								t.close()
 						if nca.header.getRightsId() == 0:
 							nca.rewind()				
 							encKeyBlock = nca.header.getKeyBlock()	
 							if str(keypatch) != "False":
+								t = tqdm(total=False, unit='B', unit_scale=False, leave=False)									
 								if keypatch < nca.header.getCryptoType2():
-									encKeyBlock,crypto1,crypto2=squirrelXCI.get_new_cryptoblock(nca, keypatch,encKeyBlock,t)					
+									encKeyBlock,crypto1,crypto2=squirrelXCI.get_new_cryptoblock(squirrelXCI,nca,keypatch,encKeyBlock,t)					
+								t.close()									
 						nca.rewind()					
 						i=0				
 						newheader=xci.get_newheader(nca,encKeyBlock,crypto1,crypto2,hcrypto,gc_flag)		
@@ -722,11 +736,182 @@ def gen_xci_parts(filepath,target_cnmt=None,cachefolder=None,keypatch=False,expo
 			if i==2:	
 				csvfile.write("{}|{}|{}|{}|{}|{}".format(1,filepath,( os.path.getsize(filepath)),targetsize,off1,off2))		
 				i+=1
-	g=os.path.basename(filepath) 			
-	g0=[pos for pos, char in enumerate(g) if char == '[']
-	g0=(g[0:g0[0]]).strip()			
-	nspname="test.nsp"
+	nspname="test.nsp"				
 	try:
+		g=os.path.basename(filepath) 			
+		g0=[pos for pos, char in enumerate(g) if char == '[']
+		g0=(g[0:g0[0]]).strip()			
 		nspname=f"{g0} [{titleid}] [v{titleversion}] [{ctype}].nsp"
 	except:pass
 	return nspname
+
+def gen_xci_parts_spec1(filepath,target_cnmt=None,cachefolder=None,keypatch=False):
+	if keypatch!=False:
+		try:
+			keypatch=int(keypatch)
+		except:	keypatch=False
+	if cachefolder==None:
+		cachefolder=os.path.join(ztools_dir, '_mtp_cache_')	
+	if not os.path.exists(cachefolder):
+		os.makedirs(cachefolder)
+	else:
+		for f in os.listdir(cachefolder):
+			fp = os.path.join(cachefolder, f)
+			try:
+				shutil.rmtree(fp)
+			except OSError:
+				os.remove(fp)
+	files_list=sq_tools.ret_xci_offsets(filepath)
+	files=list();filesizes=list()
+	fplist=list()
+	for k in range(len(files_list)):
+		entry=files_list[k]
+		fplist.append(entry[0])
+	if target_cnmt==None:	
+		for i in range(len(files_list)):			
+			entry=files_list[i]
+			cnmtfile=entry[0]	
+			if cnmtfile.endswith('.cnmt.nca'):
+				target_cnmt=cnmtfile
+				break				
+	for i in range(len(files_list)):
+		entry=files_list[i]
+		cnmtfile=entry[0]
+		if cnmtfile.endswith('.cnmt.nca') and target_cnmt==cnmtfile:
+			f=squirrelXCI(filepath)
+			titleid,titleversion,base_ID,keygeneration,rightsId,RSV,RGV,ctype,metasdkversion,exesdkversion,hasHtmlManual,Installedsize,DeltaSize,ncadata=f.get_data_from_cnmt(cnmtfile)
+			f.flush()
+			f.close()
+			for j in range(len(ncadata)):
+				row=ncadata[j]
+				# print(row)
+				if row['NCAtype']!='Meta' and row['NCAtype']!='Program':
+					test1=str(row['NcaId'])+'.nca';test2=str(row['NcaId'])+'.ncz'
+					if test1 in fplist:
+						files.append(str(row['NcaId'])+'.nca')
+						filesizes.append(int(row['Size']))
+					elif test2 in fplist:	
+						files.append(str(row['NcaId'])+'.ncz')
+						for k in range(len(files_list)):
+							entry=files_list[k]
+							if entry[0]==test2:				
+								filesizes.append(int(entry[3]))	
+								break				
+			for j in range(len(ncadata)):
+				row=ncadata[j]						
+				if row['NCAtype']=='Meta':
+					# print(str(row['NcaId'])+'.cnmt.nca')
+					files.append(str(row['NcaId'])+'.cnmt.nca')
+					filesizes.append(int(row['Size']))	
+			for j in range(len(ncadata)):
+				row=ncadata[j]
+				# print(row)
+				if row['NCAtype']=='Program':
+					test1=str(row['NcaId'])+'.nca';test2=str(row['NcaId'])+'.ncz'
+					if test1 in fplist:
+						files.append(str(row['NcaId'])+'.nca')
+						filesizes.append(int(row['Size']))
+					elif test2 in fplist:	
+						files.append(str(row['NcaId'])+'.ncz')
+						for k in range(len(files_list)):
+							entry=files_list[k]
+							if entry[0]==test2:				
+								filesizes.append(int(entry[3]))	
+								break				
+			break										
+	f.flush()
+	f.close()						
+	outheader = sq_tools.gen_nsp_header(files,filesizes)	
+	properheadsize=len(outheader)
+	# print(properheadsize)
+	# print(bucketsize)
+	i=0;sum=properheadsize;
+	xci=squirrelXCI(filepath)
+	outfile=os.path.join(cachefolder, "0")
+	outf = open(outfile, 'w+b')		
+	outf.write(outheader)	
+	written=0
+	for nspF in xci.hfs0:
+		if str(nspF._path)=="secure":
+			for ticket in nspF:			
+				if type(ticket) == Ticket:
+					size=ticket.size			
+					size_pr=sq_tools.getSize(size)			
+					filename =  str(ticket._path)
+					tik=filename[:-20]				
+					if str(tik).upper() == str(titleid).upper():
+						titleKey = ticket.getTitleKeyBlock().to_bytes(16, byteorder='big')
+						break	
+	for fi in files:
+		for nspF in xci.hfs0:	
+			if str(nspF._path)=="secure":
+				for nca in nspF:					
+					if nca._path==fi:
+						nca=Nca(nca)
+						crypto1=nca.header.getCryptoType()
+						crypto2=nca.header.getCryptoType2()	
+						if crypto2>crypto1:
+							masterKeyRev=crypto2
+						if crypto2<=crypto1:	
+							masterKeyRev=crypto1									
+						crypto = aes128.AESECB(Keys.keyAreaKey(Keys.getMasterKeyIndex(masterKeyRev), nca.header.keyIndex))
+						hcrypto = aes128.AESXTS(uhx(Keys.get('header_key')))	
+						gc_flag='00'*0x01					
+						crypto1=nca.header.getCryptoType()
+						crypto2=nca.header.getCryptoType2()					
+						if nca.header.getRightsId() != 0:					
+							nca.rewind()	
+							if crypto2>crypto1:
+								masterKeyRev=crypto2
+							if crypto2<=crypto1:	
+								masterKeyRev=crypto1								
+							titleKeyDec = Keys.decryptTitleKey(titleKey, Keys.getMasterKeyIndex(int(masterKeyRev)))							
+							encKeyBlock = crypto.encrypt(titleKeyDec * 4)
+							if str(keypatch) != "False":
+								t = tqdm(total=False, unit='B', unit_scale=False, leave=False)	
+								if keypatch < nca.header.getCryptoType2():
+									encKeyBlock,crypto1,crypto2=squirrelXCI.get_new_cryptoblock(squirrelXCI,nca,keypatch,encKeyBlock,t)	
+								t.close()
+						if nca.header.getRightsId() == 0:
+							nca.rewind()											
+							encKeyBlock = nca.header.getKeyBlock()	
+							if str(keypatch) != "False":
+								t = tqdm(total=False, unit='B', unit_scale=False, leave=False)								
+								if keypatch < nca.header.getCryptoType2():
+									encKeyBlock,crypto1,crypto2=squirrelXCI.get_new_cryptoblock(squirrelXCI,nca,keypatch,encKeyBlock,t)	
+								t.close()									
+						nca.rewind()					
+						i=0				
+						newheader=xci.get_newheader(nca,encKeyBlock,crypto1,crypto2,hcrypto,gc_flag)	
+						outf.write(newheader)
+						written+=len(newheader)
+						nca.seek(0xC00)	
+						break					
+					else:pass					
+	xci.flush()
+	xci.close()							
+	outf.flush()							
+	outf.close()		
+	tfile=os.path.join(cachefolder, "files.csv")	
+	with open(tfile,'w') as csvfile:	
+		csvfile.write("{}|{}|{}|{}|{}|{}\n".format("step","filepath","size","targetsize","off1","off2"))	
+		csvfile.write("{}|{}|{}|{}|{}|{}\n".format(0,outfile,properheadsize+written,properheadsize,0,properheadsize))	
+		k=0;l=0	
+		for fi in files:
+			for j in files_list:
+				if j[0]==fi:	
+					csvfile.write("{}|{}|{}|{}|{}|{}\n".format(k+1,outfile,properheadsize+written,0xC00,(properheadsize+l*0xC00),(properheadsize+(l*0xC00)+0xC00)))	
+					off1=j[1]+0xC00
+					off2=j[2]
+					targetsize=j[3]-0xC00				
+					csvfile.write("{}|{}|{}|{}|{}|{}\n".format(k+2,filepath,(os.path.getsize(filepath)),targetsize,off1,off2))	
+					break
+			k+=2;l+=1	
+	nspname="test.nsp"				
+	try:
+		g=os.path.basename(filepath) 			
+		g0=[pos for pos, char in enumerate(g) if char == '[']
+		g0=(g[0:g0[0]]).strip()			
+		nspname=f"{g0} [{titleid}] [v{titleversion}] [{ctype}].nsp"
+	except:pass
+	return nspname
